requirements.txt
streamlit
opencv-python
numpy
Pillow
matplotlib
scipy
